Add files via upload
updated the file, removing redundant lines and adding Mstar and Rstar histograms and R-M plot. Also filtered out stars that have an absolute K magnitude (M_Ks) outside the range for the relations in Mann et al. (2015). As it turns out, quite a number of stars do have a M_Ks outside this range, which is [4.5, 9.5).