Add files via upload
forgot to include that the G - Ks relation only holds for G<13. Updated this, the result is a slight reduction in the number of stars in the csv file. To be more precise, the K mag histogram in queryHistogram.py now counts 238 stars with a absolute K mag (M_Ks) between 4.5 and 9.5.